requirements.txt
azure-cosmos==4.2.0
Flask==2.1.1
Pandas==1.4.1